Move PnP client side API to new location
