+++ Core/OfficeDevPnP.Core/Sites/Dummy.txt
@@ -1,0 +1,1 @@
+﻿Remove when adding classes in this folder
